Move Convex setup to webcontainer boot
