+++ app/lib/runtime/action-runner.ts
@@ -6,8 +6,7 @@ import { createScopedLogger } from '~/utils/logger';
 import { unreachable } from '~/utils/unreachable';
 import type { ActionCallbackData } from './message-parser';
 import type { BoltShell } from '~/utils/shell';
-import { convexStore, waitForConvexProjectConnection } from '~/lib/stores/convex';
-import { initializeConvexAuth } from '~/lib/convexAuth';
+import { waitForConvexProjectConnection } from '~/lib/stores/convex';
 import type { ToolInvocation } from 'ai';
 import { withResolvers } from '~/utils/promises';
 import { BackupStack, editor, editorToolParameters } from './editorTool';
@@ -210,7 +209,7 @@ export class ActionRunner {
           return;
         }
         case 'convex': {
-          await this.#runConvexAction(actionId, action);
+          logger.error("Convex action is not supported anymore. Use tool calls instead.");
           break;
         }
         case "toolUse": {
@@ -398,50 +397,6 @@ export class ActionRunner {
     };
   }
 
-  async #runConvexAction(actionId: string, action: ActionState) {
-    if (action.type !== 'convex') {
-      unreachable('Expected convex action');
-    }
-
-    const convexLogger = createScopedLogger('ActionRunner:Convex');
-    const webcontainer = await this.#webcontainer;
-
-    await waitForConvexProjectConnection();
-
-    await this.#setupConvexEnvVars();
-
-    const updateAction = this.updateAction.bind(this);
-
-    // Run convex dev --once
-
-    const process = await webcontainer.spawn('npx', ['convex', 'dev', '--once'], {
-      env: { npm_config_yes: true },
-    });
-
-    action.abortSignal.addEventListener('abort', () => {
-      process.kill();
-    });
-
-    let fullOutput = '';
-    process.output.pipeTo(
-      new WritableStream({
-        write(data) {
-          fullOutput += data;
-          convexLogger.debug('Convex output', data);
-          updateAction(actionId, { output: fullOutput });
-        },
-      }),
-    );
-
-    const exitCode = await process.exit;
-
-    if (exitCode !== 0) {
-      throw new ActionCommandError('Convex Dev Failed', fullOutput || 'No Output Available');
-    }
-
-    convexLogger.info(`Convex process terminated with code ${exitCode}`);
-  }
-
   async #runToolUseAction(actionId: string, action: ActionState) {
     const parsed: ToolInvocation = JSON.parse(action.content);
     if (parsed.state === "result") {
@@ -497,56 +452,6 @@ export class ActionRunner {
       }
       resolvers.resolve(message);
       throw e;
-    }
-  }
-
-  async #setupConvexEnvVars() {
-    const webcontainer = await this.#webcontainer;
-
-    const convexProject = convexStore.get();
-
-    if (!convexProject) {
-      throw new Error('No Convex project found');
-    }
-
-    const { token } = convexProject;
-
-    const envFilePath = '.env.local';
-    const envVarName = 'CONVEX_DEPLOY_KEY';
-    const envVarLine = `${envVarName}=${token}\n`;
-
-    let content: string | null = null;
-    try {
-      content = await webcontainer.fs.readFile(envFilePath, 'utf-8');
-    } catch (err: any) {
-      if (!err.toString().includes('ENOENT')) {
-        throw err;
-      }
-    }
-    if (content === null) {
-      // Create the file if it doesn't exist
-      await webcontainer.fs.writeFile(envFilePath, envVarLine);
-      logger.debug('Created .env.local with Convex token');
-    } else {
-      const lines = content.split('\n');
-
-      // Check if the env var already exists
-      const envVarExists = lines.some((line) => line.startsWith(`${envVarName}=`));
-
-      if (!envVarExists) {
-        // Add the env var to the end of the file
-        const newContent = content.endsWith('\n') ? `${content}${envVarLine}` : `${content}\n${envVarLine}`;
-        await webcontainer.fs.writeFile(envFilePath, newContent);
-        logger.debug('Added Convex token to .env.local');
-      } else {
-        logger.debug('Convex token already exists in .env.local');
-      }
-    }
-
-    try {
-      await initializeConvexAuth(convexProject);
-    } catch (error) {
-      logger.error('Failed to initialize Convex Auth', error);
     }
   }
 }
